renames files to reflect class name changes
