+++ lib/booking_system.rb
