--- lib/hotel.rb
@@ -1,8 +1,0 @@
-class Hotel
-  attr_reader :hotel
-
-  def initialize
-    @hotel
-  end
-
-end
