+++ spec/booking_system_spec.rb
