--- spec/hotel_spec.rb
@@ -1,8 +1,0 @@
-# require_relative 'spec_helper'
-# # require_relative '../lib/hotel'
-#
-# describe 'Hotel class' do
-#   it 'can instantiate a new Hotel object' do
-#     expect(Hotel.new).must_be_instance_of Hotel
-#   end
-# end
