+++ spec/spec_helper.rb
@@ -16,6 +16,6 @@ Minitest::Reporters.use! Minitest::Reporters::SpecReporter.new
 # require_relative '../lib/hotel'
 
 require_relative '../lib/date_range'
-require_relative '../lib/hotel_booker'
+require_relative '../lib/booking_system'
 require_relative '../lib/reservation'
-require_relative '../lib/room_block'
+require_relative '../lib/block'
